refactor into smaller components
