+++ src/css/leftpanel.scss
@@ -709,7 +709,7 @@
     padding-right: 0.3rem;
   }
   .imagery-secondary-filters {
-    margin-top: 10px;
+    margin-top: 15px;
     border-bottom: #e5e5df 0.5px solid;
     display: grid;
     grid-template-columns: 1fr auto;
